Rename `ContextParallelLoadBalancer` to `RingAttention*`
