+++ src/olmo_core/distributed/parallel/__init__.py
@@ -8,13 +8,7 @@ from olmo_core.distributed.utils import get_num_nodes, get_world_size
 from olmo_core.exceptions import OLMoConfigurationError
 from olmo_core.utils import get_default_device
 
-from .context_parallel import (
-    ContextParallelConfig,
-    ContextParallelLlama3LoadBalancer,
-    ContextParallelLoadBalancer,
-    ContextParallelLoadBalancerType,
-    ContextParallelZigZagLoadBalancer,
-)
+from .context_parallel import ContextParallelConfig
 from .data_parallel import DataParallelConfig, DataParallelType, DPMeshDimName
 from .expert_parallel import ExpertParallelConfig
 from .pipeline_parallel import (
@@ -45,10 +39,6 @@ __all__ = [
     "PipelineScheduleType",
     "PipelineSchedule",
     "ContextParallelConfig",
-    "ContextParallelLoadBalancerType",
-    "ContextParallelLoadBalancer",
-    "ContextParallelZigZagLoadBalancer",
-    "ContextParallelLlama3LoadBalancer",
 ]
 
 log = logging.getLogger(__name__)
