+++ src/test/nn/attention_test.py
@@ -8,6 +8,7 @@ from olmo_core.nn.attention import (
     AttentionConfig,
     AttentionType,
     FusedAttention,
+    RingAttentionZigZagLoadBalancer,
 )
 from olmo_core.nn.layer_norm import LayerNormConfig
 from olmo_core.nn.rope import RoPEConfig, RoPEType
@@ -210,3 +211,100 @@ def test_attention_buidler_config(attn_config: AttentionConfig):
     # Make sure the estimated number of params matches the actual number of params.
     n_params = sum(p.numel() for p in attn.parameters())
     assert attn_config.num_params(d_model) == n_params
+
+
+def _get_lb(rank: int, world_size: int) -> RingAttentionZigZagLoadBalancer:
+    return RingAttentionZigZagLoadBalancer(cp_rank=rank, cp_world_size=world_size)
+
+
+def test_zig_zag_load_balancer_padding():
+    x, padding_added = _get_lb(0, 4).pad(torch.tensor([0, 1, 2, 3, 4, 5]).unsqueeze(0), 1, -1)
+    assert x.tolist() == [[0, 1, 2, 3, 4, 5, -1, -1]]
+    assert padding_added == 2
+
+
+def test_zig_zag_load_balancer_shard():
+    x = torch.tensor([0, 1, 2, 3, 4, 5, 6, 7]).unsqueeze(0)
+    assert _get_lb(0, 4).batch_shard(inputs=[x], seq_dims=[1])[0].tolist() == [
+        [
+            0,
+            7,
+        ]
+    ]
+    assert _get_lb(3, 4).batch_shard(inputs=[x], seq_dims=[1])[0].tolist() == [
+        [
+            3,
+            4,
+        ]
+    ]
+
+
+def test_zig_zag_load_balancer_shard_with_padding():
+    x = torch.tensor([0, 1, 2, 3, 4, 5]).unsqueeze(0)
+    assert _get_lb(0, 4).batch_shard(inputs=[x], seq_dims=[1], pad_values=[-1])[0].tolist() == [
+        [
+            0,
+            -1,
+        ]
+    ]
+    assert _get_lb(3, 4).batch_shard(inputs=[x], seq_dims=[1], pad_values=[-1])[0].tolist() == [
+        [
+            3,
+            4,
+        ]
+    ]
+
+
+def test_zig_zag_load_balancer_shard_by_document():
+    x = torch.tensor(list(range(12))).unsqueeze(0)
+    cu_doc_lens = torch.tensor([0, 8, 12])
+
+    assert _get_lb(0, 2).batch_shard_by_document(inputs=[x], seq_dims=[1], cu_doc_lens=cu_doc_lens)[
+        0
+    ][0].tolist() == [
+        [
+            0,
+            1,
+            6,
+            7,
+            8,
+            11,
+        ]
+    ]
+
+    assert _get_lb(1, 2).batch_shard_by_document(inputs=[x], seq_dims=[1], cu_doc_lens=cu_doc_lens)[
+        0
+    ][0].tolist() == [
+        [
+            2,
+            3,
+            4,
+            5,
+            9,
+            10,
+        ]
+    ]
+
+
+def test_zig_zag_load_balancer_shard_by_document_with_padding():
+    x = torch.tensor(list(range(12))).unsqueeze(0)
+    cu_doc_lens = torch.tensor([0, 7, 10])
+
+    res, opts = _get_lb(0, 2).batch_shard_by_document(
+        inputs=[x],
+        seq_dims=[1],
+        cu_doc_lens=cu_doc_lens,
+        pad_values=[-1],
+    )
+    new_doc_lens = opts["cu_doc_lens"]
+    assert new_doc_lens.tolist() == [0, 4, 6]
+    assert res[0].tolist() == [
+        [
+            0,
+            1,
+            6,
+            -1,
+            7,
+            -1,
+        ]
+    ]
